[CLN] base,*: drop BaseUsersCommon
This common is not much used, nor useful.
All it did was create two new users out of the box, but
it can be done easily with the tooling from the standard
BaseCommon it had as parent.

As the BaseCommon is expected to be the base for the new
standard test architecture, and we want to keep it as simple
as possible, removing this common is better.
